Previous Square Overwite
This guards against players overwriting a previous move on the game board. Now, because of `def SqaureIsTaken (chk):` we can now check if squares contain either a X or O and If this statement is true, then ask user for a new and valid input. Once a valid input has been entered, then it will proceed in the main script to update the board.